==> Code/2_alpha_beta_diersity.Rmd <==
---
title: "2_alpha_beta_diersity"
author: "Pedro"
date: "8/9/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

# 0 - Load libraries & data

```{r}

#load necessary libraries
library(phyloseq)
library(ggplot2)
library(vegan)
library(EcolUtils)
library(microbiome)
library(car)
library(agricolae)

#load the phyloseq objects containing all microbiome data
load("./Code/phyloseq_objects.RData")

```


# 1 - Check beta diversity in a NMDS ordination
Some clustering and treatment effects are evident!
```{r}

# Let's make a Non-Metric Multidimensional Scaling (NMDS) of all our samples based on CSS normalization
nmds   <- phyloseq::ordinate(physeq_goodSamples_CSS, 
                             method ="PCoA", # this method has few assumptions and readly accepts different data structures
                             distance="bray", # bray-curtis distance is suitable for sparse data - such as the zero-inflated microbiome data we have
                             try=200, # number of iterations
                             autotransform=TRUE) # automatically transforms your data, if needed. reduces weight of outliers
                              #weakties = FALSE prevests tress from colapsing into zero

plot_ordination(physeq = physeq_goodSamples_CSS, 
                ordination = nmds, 
                color= "Treatment")+ 
  theme_classic() +
  theme(plot.title = element_text(size = 10, face = "bold", hjust = 0.5)) +
  theme(legend.position="right")




```

# 2 - check beta dispersion
beta dispersion is like multivatite homogeniety of variances. just like you can't compare properly very small error bars with very large error bars, you should check fi the spread of the data clouds is similar per treatment
```{r}




#calculates beta dispersions of the treatment effects
beta_disper<- betadisper(d = phyloseq::distance(t(otu_table(physeq_goodSamples_CSS)), method="bray"),
                         group = sample_data(physeq_goodSamples_CSS)[["Treatment"]],
                         type = "centroid", sqrt.dist = TRUE) 


# check if ebta dispersions are significantly differente
set.seed(100)
beta_disper_calc<-anova(beta_disper, permutations = 9999)
betadisp_tukey<-TukeyHSD(beta_disper, ordered)
scores(beta_disper)
boxplot(beta_disper)

# saves the p value of the dispersion test
p_dispersion<-beta_disper_calc[[5]][1] 

#plots a spdier plot (PCoA) with centroids
beta_diversity<-plot(beta_disper,
     main = "Beta Dispersion",
     sub = p_dispersion,
     xlab="p value for homogeniety of multivariate variances:")


#export as pdf
ggsave(beta_diversity, file = "./Results/1_Beta_Dipersion_PCoA.pdf", 
       width = 15, 
       height = 15, 
       units = "cm"  )

# aditional betadispersion option
sample_data(physeq_goodSamples_CSS)$treatment_type<-
  c(rep(x = "Insects", 7), 
    rep(x = "Controls", 15), 
    rep(x = "Insects", 13), 
    rep(x = "Controls", 5))



#calculates beta dispersions of the treatment effects
beta_disper2<- betadisper(d = phyloseq::distance(t(otu_table(physeq_goodSamples_CSS)), method="bray"),
                         group = sample_data(physeq_goodSamples_CSS)[["treatment_type"]],
                         type = "centroid", sqrt.dist = TRUE) 


# check if ebta dispersions are significantly differente
set.seed(100)
beta_disper_calc<-anova(beta_disper2, permutations = 9999)
betadisp_tukey<-TukeyHSD(beta_disper2)
scores(beta_disper2)
boxplot(beta_disper2)






```

## 2.1 -  export beta dispersion statistics
```{r}
boxplot(beta_disper)

library(flextable)

betadisp_anova_ft<-flextable(round(beta_disper_calc,digits = 3))

betadisp_tukey_export<-round(as.data.frame(betadisp_tukey$group), digits = 3)
betadisp_tukey_export[,3]<-rownames(betadisp_tukey_export)
betadisp_tukey_ft<-flextable(betadisp_tukey_export[3:4])

save_as_docx(  "Beta Dispersion ANOVA test" = betadisp_anova_ft, 
               "Tukey HSD for Beta Dispersion" = betadisp_tukey_ft, 
                path = "./Results/betadisp_table.docx")

```

## 2.2 - compare beta dispersion of exuviae treatments VS other treatments


```{r}
# add new metadata column, separating excuviae treatments from non-excuvie treatments

physeq_goodSamples_CSS@sam_data$Exuviae_yes_no<-
ifelse(test = physeq_goodSamples_CSS@sam_data$Treatment %in% c("BSF", "MW", "HC"),
       yes = "Exuviae treatments",
       no = "Control treatments")



#calculates beta dispersions of the treatment effects
beta_disper<- betadisper(d = phyloseq::distance(t(otu_table(physeq_goodSamples_CSS)), method="bray"),
                         group = sample_data(physeq_goodSamples_CSS)[["Exuviae_yes_no"]],
                         type = "centroid", 
                         sqrt.dist = TRUE)


# check if ebta dispersions are significantly differente
beta_disper_calc<-anova(beta_disper, permutations = 999)

#visualize plot
boxplot(beta_disper)

```


# 3 - run PERMANOVA and post-hoc test
We see significat treatment effects, with clear paiwise distinctions

```{r}
# Running the permanova with vegan::adonis2() on a single phyloseq object is very simple
metadata<-as(sample_data(physeq_goodSamples_CSS),"data.frame")
#let's first define permitation design with permute::how()
perm2<-how(# define blocks
           nperm=999)# n of permutations

# now let's calculate the permanova
adonis2(phyloseq::distance(t(otu_table(physeq_goodSamples_CSS)), method="bray") # this is your distance matrix of OTU abundances 
        ~ Treatment, # this is your model
        permutations=how(nperm=999), # we jsut defined the permutation settings above, you place them here
        data = metadata) # metadata for the community matrix


# post-hoc
adonis.pair(dist.mat= phyloseq::distance(otu_table(physeq_goodSamples_CSS), method="bray"),
              Factor= as.factor(as(phyloseq::sample_data(physeq_goodSamples_CSS),"data.frame")$Treatment))
```


# 4 - Alpha diversity calculations & plot


```{r}


#Calculate richness for root and soil
calculated_diversity<-microbiome::diversity(physeq_goodSamples_rarefied)

n_sequences<-sample_sums(physeq_goodSamples_rarefied)
calculated_diversity$n_sequences<-n_sequences

#### add diversity metrics to mapping file of phyloseq objects
# we do this so we can perform anovas, acess metadat, make nicer plots, etc
merg_to_ps<-sample_data(calculated_diversity) # makes the diversity calculations  sample_data for phyloseq oject...
calculated_diversity_df<-as(sample_data(merge_phyloseq(physeq_goodSamples_rarefied,merg_to_ps)),"data.frame") # forces sample data of updated phyloseq object into a dataframe

#Lets see this on a simple boxplot for Shannon diversity index
fisher_div<-ggplot(calculated_diversity_df, aes(x =Treatment, y = fisher, fill =Treatment ))+
  geom_boxplot()+
  theme_bw()+
  labs(title = "Fisher diversity indexes")


#export as pdf
ggsave(fisher_div, file = "./Results/3_fisher_diversity.pdf", 
       width = 15, 
       height = 15, 
       units = "cm"  )



```



# 5 - alpha diversity tests
Shannon diversity has a bit of heterogeniety per treatent:  p= 0.051. Fisher diversity instead has p=0.90 and the pattern in teh plot is the same. stats and post-hoc of shannon and fisher are similar.  let's use fisher diversities here due to lower heterogeniety


```{r}
# chech homogeniety of variances
leveneTest((fisher) ~ Treatment, data = calculated_diversity_df) 


# anova test
anova_fisher<-aov(fisher ~ Treatment, data = calculated_diversity_df)

# get p values
summary(anova_fisher)

# post-hoc, both Tukey and Least Significat Difference
LSD.test(y = anova_fisher, "Treatment", group=TRUE, console=TRUE, p.adj= "fdr")
TukeyHSD(anova_fisher)

```
