==> Code/3_differential_abundance.Rmd <==
---
title: "3_differential_abundance"
author: "Pedro"
date: "8/9/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

# 0 - load libraries and data
```{r}

#load necessary libraries
library(phyloseq)
library(ggplot2)
library(DESeq2)
library(pheatmap)
library(viridisLite)
library(metacoder)


#load the phyloseq objects containing all microbiome data
load("./Code/phyloseq_objects.RData")

```



# 1 - create deseq object and calculate differential abudnaces
```{r}


# make list of deseqed phyloseq objects
deseq_obj<- phyloseq_to_deseq2(physeq_goodSamples, ~ Treatment)
deseq_obj<- DESeq(deseq_obj, test="Wald", fitType="parametric")



```



# 2 - define pairwise comparisons, extract  with custom function
```{r}

physeq_goodSamples@sam_data$Treatment

# make a list of the pairwise comparisons you want to run according the treatment ; 
#  positive fold change: higher in fist factor (treatment) / negative fold change: higher in second factor(control)
pairwise_comparison<-list(
c("Treatment", "BSF", "C"),
c("Treatment", "BSF", "CHT"), 
c("Treatment", "BSF", "HC"),
c("Treatment", "BSF", "MW"),
c("Treatment", "BSF", "OF"),
c("Treatment", "C", "CHT"),
c("Treatment", "C", "HC"),
c("Treatment", "C", "MW"),
c("Treatment", "C", "OF"),
c("Treatment", "CHT", "HC"),
c("Treatment", "CHT", "MW"),
c("Treatment", "CHT", "OF"),
c("Treatment", "HC", "MW"),
c("Treatment", "HC", "OF"),
c("Treatment", "MW", "OF"))




# define function to extract pairwise comparsions p valyes, showing ASVs that are different and its taxonomeis
build_sig_tab<-function (deseq_obj, contrast_cha_vector){

  alpha<-0.05 # defines the alpha level for p adjustment to control for false discovery rate
  min_fold<-2 # defines minimum fold difference in counts to classify an OTU as differential


  list_res <- lapply(contrast_cha_vector, function(x) 
    results(deseq_obj, contrast=x, cooksCutoff = FALSE))


  list_sigtab <- lapply(list_res, function (x) 
    x[which(x$padj < alpha), ]) # cuts results based on alpha for FDR
  
   list_sigtab <- lapply(list_sigtab, function (x) 
    x[which(abs(x$log2FoldChange) > min_fold), ]) # cuts results based on minimum fold change
   
  list_sigtab <- lapply(list_sigtab, function (x) 
    x[order(x$padj),])  # sorts lists of OTUs based on p values of pairwise abundance differences
  
  
  
  
  
  
  # this silenced chuck allows you to limit the number of OTUs accepted as diferential in each comparision. note that if you define the minimum number as 1:40 (1 to 40) but you only hace 39 differentially abundannt OTUs the function will crash and you will not have these results. just remove the hashtags below to limt to 3 top differentially abundant OTUs
  
    # max_OTU_n<-1:3 # defines the maximum number of OTUs to include in each pairwise coparison
    # list_sigtab <- lapply(list_sigtab, function (x) 
    # x[max_OTU_n,])  # cuts results based on maximum numer of OTUs counted as differential
  
  
  
  
  
  
    
  list_sigtab <- lapply(list_sigtab, function (x) 
    cbind(as(x, "data.frame"), as(tax_table(physeq_goodSamples)[rownames(x), ], "matrix"))) # here we only colect the taxa names from the original, unsplit, phyloseq object
  return(list_sigtab)
  }

# run custom function
sigtab_results<-build_sig_tab(deseq_obj = deseq_obj,contrast_cha_vector = pairwise_comparison)

# extract ASV names of the differentially abundant ASVs
differentialy_abundant_ASVs<-lapply(sigtab_results, function (x) rownames(x))%>% # extract the rownames from the list, then....
unlist(use.names=FALSE)%>% # flatten the rowname (OTU ID) list, ignoring the names of the lists, then...
  unique() # remove duplications of the same OTUs




# extract Orders from teh sigtabs, showing how often an order is represented
Order_representation<-lapply(sigtab_results, function (x) x$Order)%>% # extract the rownames from the list, then....
unlist(use.names=FALSE)%>% # flatten the rowname (OTU ID) list, ignoring the names of the lists, then...
  table()%>% # put it on a table, showing ORders that are more often represented
sort()# sort that table


#extarct orders of the 62 uniquely differential ASVs
table(tax_table(prune_taxa(differentialy_abundant_ASVs, physeq_goodSamples))[,4])%>% 
  sort()
```

# 3 - plot bi-clsuter heatmap
```{r}

# finally, build the heatmap
# note that the log change calculations are based on the full, unsplit and unlisted deseq2 object (that is, our complete dataset)
ntb <- normTransform(deseq_obj) # defaults to log2(x+1)
log2.norm.countsb <- assay(ntb)[differentialy_abundant_ASVs, ] #these colnames should match rownames(df)
df <- as.data.frame(colData(deseq_obj)[,c("Treatment")])
row.names(df)<-row.names(as.data.frame(colData(deseq_obj)))
df_row <- as.data.frame((tax_table(physeq_goodSamples))[,c("Order")]) # here you can change to Class, ORder, etc...S
heatmap_output<-pheatmap::pheatmap(log2.norm.countsb, color=viridis(24), annotation_col=df, annotation_row=df_row, main="log2(counts + 1)", 
                   scale="none", fontsize_row = 3, fontsize_col = 8)


#export as pdf
ggsave(beta_diversity, file = "./Results/2_deseq_order.pdf", 
       width = 20, 
       height = 20, 
       units = "cm"  )



```
