==> Code/5_network_analysis.Rmd <==
---
title: "5_network_analysis"
author: "Pedro Beschoren"
date: "2022-12-22"
output: html_document
editor_options: 
  chunk_output_type: console
---



# load data, libraries
```{r}


library("BDgraph") # check overlap of bulk soil core ASVs
library("ggplot2")
library("tidyr")
library("plyr")
library("dplyr")
library("parallel")
library("phyloseq")
library("tibble")
library("igraph")
library("purrr")
library("ggrepel")
library("tibble")
library(ggpubr)

#check memory and cores
memory.size()
detectCores()

# load custom netowrk functions
source("./Code/Network_custom_functions.R")

# load data
load("./Code/phyloseq_objects.RData")

load(file = "./Results/model_borutized.RData")



#this is your main ps object
physeq_goodSamples_rarefied

# these are teh ASVs modeled as important by boruta, and cross-validated in caret
model_borutized$coefnames

# make a new ps object, that only has AVSs classified as important by boruta
network_input_ps<-prune_taxa(model_borutized$coefnames,physeq_goodSamples_rarefied)

```




## prepare and check network input
matrix_input and metadata_inputwill be used as input to bdgraph.dw. here we check some of their structure

```{r echo=TRUE}



# transform phyloseq objects otus into matrixes
matrix_input<-t(as(otu_table(network_input_ps), "matrix"))

# transform phyloseq objects metadata into df
metadata_input<-as(sample_data(network_input_ps), "data.frame")

# force some metadata into  factor
metadata_input$Treatment<-as.factor(metadata_input$Treatment)


# what is the percentage of zeros in this matrix?
sum(colSums(matrix_input==0))/length(matrix_input)*100

colSums(matrix_input)%>%sort
rowSums(matrix_input)%>%sort



```


```{r}
            
#calculate network, removing spurious edges caused by treament metadata

# this tep was performed on WUR's HPC annuna, consuming 1.1 core and 730Gb memory for 18h
# BD_CSS_out_4.5M<-bdgraph.dw(data = matrix_input, 
#                   x = metadata_input,
#                   algorithm = "bdmcmc",
#                   formula = y ~  Treatment,
#                   cores = "all",
#                   iter_bdw = 1500000,
#                   iter = 300000,
#                   save = TRUE)

# save output externallu
# save(BD_CSS_out_4.5M, file = "./Data/BD_CSS_out_4.5M.RData")





```



# load graphs calculated on HPC
```{r}
# load bd graph object
load(file = "./Data/BD_CSS_out_4.5M.RData")

#this is your network
BD_CSS_out_4.5M
  

  
#plot edge stability, then save as jpeg
jpeg(file="./Results/coda4.5M.jpeg")
plotcoda(BD_CSS_out_4.5M)
dev.off()

#plot graph summary, then save as jpeg
jpeg(file="./Results/summary4.5M.jpeg")
summary.bdgraph(BD_CSS_out_4.5M)
dev.off()

```

# Check p link  statistics
```{r}


# check bdgrph plink distribuion
plink_vector<- plinks(BD_CSS_out_4.5M)[upper.tri(plinks(BD_CSS_out_4.5M))]
hist(plink_vector)
summary(plink_vector)

# network with plinks on teh 3rd quartile
summary(plink_vector)[[5]]
```
# create list of different plink cutoffs
```{r}
# check different cutoffs

bd_cutoff_07<-BDgraph::select(BD_CSS_out_4.5M,  cut = 0.7, vis = TRUE)  #same  as random
bd_cutoff_06<-BDgraph::select(BD_CSS_out_4.5M,  cut = 0.6, vis = TRUE)  #same  as random
bd_cutoff_md2sd<-BDgraph::select(BD_CSS_out_4.5M,  cut = median(plink_vector)+2*sd(plink_vector), vis = TRUE)  #same  as random
bd_cutoff_md1sd<-BDgraph::select(BD_CSS_out_4.5M,  cut = median(plink_vector)+1*sd(plink_vector), vis = TRUE)  #differs from ramdom
bd_cutoff_md<-BDgraph::select(BD_CSS_out_4.5M,  cut = median(plink_vector), vis = TRUE) # differs from ramdom
bd_cutoff_625<-BDgraph::select(BD_CSS_out_4.5M,  cut = 0.625, vis = TRUE) # p08
bd_cutoff_65<-BDgraph::select(BD_CSS_out_4.5M,  cut = 0.65, vis = TRUE) # p08
bd_cutoff_675<-BDgraph::select(BD_CSS_out_4.5M,  cut = 0.675, vis = TRUE) # p08


#save as list
bd_list<-list( "bd_cutoff_05" = summary.bdgraph(BD_CSS_out_4.5M)$selected_g,
               "bd_cutoff_06"= bd_cutoff_06, 
               "bd_cutoff_625"= bd_cutoff_625,
              "bd_cutoff_65" = bd_cutoff_65,
              "bd_cutoff_675" = bd_cutoff_675,
              "bd_cutoff_07" = bd_cutoff_07,
              "bd_cutoff_md2sd" = bd_cutoff_md2sd,
              "bd_cutoff_md1sd" = bd_cutoff_md1sd,
              "bd_cutoff_md" = bd_cutoff_md)






```



# import to igraph, 
```{r}
#laod igrph library
library(igraph)

# this is my bdgraph object
bd_list

#define selected graph 
selected_graph<-summary.bdgraph(BD_CSS_out_4.5M)$selected_g
selected_graph_l<-lapply(bd_list, function(x) summary.bdgraph(x)$selected_g)

#makes igraph from dbpackage
igraph_raref_obj_l<-lapply(bd_list, function (x)
  graph_from_adjacency_matrix(adjmatrix = x, mode = "undirected"))


#check size of each graph
lapply(igraph_raref_obj_l, gsize)

```




# add edge weight to igraph
```{r}
# Function to calculate partial correlations (from Paryia;  theta = k_hat from bd_object)
calculate.strength.theta <- function(theta){
  p <- ncol(theta)
  cond.cor <- matrix(NA, ncol=p, nrow=p)
  for(i in 1:nrow(theta))
  {
    for(j in 1:ncol(theta))
    {
      cond.cor[i,j] <- - theta[i,j]/ (sqrt(theta[i,i])* sqrt(theta[j,j]))
    }
  }
  rownames(cond.cor) <- colnames(theta)
  colnames(cond.cor) <- colnames(theta)
  return(cond.cor)
}

# define a funciton to add bd_graph weights to igraph objects
add_edge_weight_to_igraph<-function(original_bd_obj, bd_obj_l, igraph_obj_l){
  # this function will add edged weights to the igraph objects
    # original_bd_obj = a bdgraph object, it MUST contain a $K_hat data layer
    # bd_obj_l = a list of bd_graph objects, such as those with different plink cutoffs,
    # igraph_obj_l= a list of igraph objects, geerated with the graph_from_adjacency_matrix)() function
      # we define the eddge weights with paryia's function, targeting the K_hat, then...
      # we extract the weights of the edges that are valid at the specific bdgraph cutoof, then...
      # weights are unlisted, following a row-by-row format that fits igraph, then..
      # check if the number of igraph and bdgraph objects edges match, then...
      # save the calculated edge weights in the igraph object
  # it returns an igraph object with the weights calculated with paryia's function on bdgraph objects
  
  

#edge's weight ; NOTE: only the original bdgraph obeject has theta (K_hat)
edge_weight<-calculate.strength.theta (original_bd_obj$K_hat)


# define weight edges of the edges selected in each plink filtering. we msut define it row by row, then unlist, to fit igraph edge listing order
edge_weight_by_row_l<-lapply(bd_obj_l, function(x){
    edge_weight_by_row<-list()
    for (i in 1:nrow(edge_weight)){
    value <-  edge_weight[i,][x[i,]>0]
    edge_weight_by_row[i]<-list(value)
    }
  return(edge_weight_by_row)
})

#unlist it
edge_weight_by_row_l<-lapply(edge_weight_by_row_l, unlist)


# check if the number of edges retained in the weight list of vectors is the same as the number of the edges in the igraph object (all sould eb TRUE)
print(" CHECK: is the number of edges retained in the weight list of vectors is the same as the number of the edges in the igraph object? all should be = TRUE!")
print(lapply(edge_weight_by_row_l, length) %in% lapply(igraph_obj_l, function (x) length(E(x))))

# save as edges in igraph
output<-mapply(function(x,y){
  E(x)$weight<-y
  return(x)
},
x = igraph_obj_l,
y = edge_weight_by_row_l)

# define output to export
return(output)
  
}


# run custm function, adding weight to teh igraph objcts
igraph_raref_obj_l<-add_edge_weight_to_igraph(original_bd_obj = BD_CSS_out_4.5M, 
                                              bd_obj_l = bd_list, 
                                              igraph_obj_l = igraph_raref_obj_l)




```

# check node prorprieties and degree histogram

```{r}

# define node degree and check degree distribution
node_prop_l<-lapply(igraph_raref_obj_l, function(x)
  Generate_node_metrics2(igraph_obj = x, physeq_obj= network_input_ps))

#define a list of histograms showing degree at each cutoff
hist_list<-lapply(node_prop_l, function(x)
  ggplot(x, aes(x=Degree))+ geom_histogram(binwidth = 1)+ xlab(NULL))

#visualize histograms
library("ggpubr")
ggarrange(hist_list[[1]],
          hist_list[[2]],
          hist_list[[3]],
          hist_list[[4]],
          hist_list[[5]],
          hist_list[[6]],
          hist_list[[7]],
          hist_list[[8]],
          hist_list[[9]],
          ncol = 1,labels = NULL, legend = "none")



```



# check slelected graph keystones
```{r}
# plink cutoff at media + 2sd, edge weight cutoff above media... provides 5/8 difference s to random and good degree distirbution
plot(igraph_raref_obj_l$bd_cutoff_md2sd)

# remove ASVs with zero edges
isolated <- which(degree(igraph_raref_obj_l$bd_cutoff_md2sd)==0)
select_igraph<- delete.vertices(igraph_raref_obj_l$bd_cutoff_md2sd, isolated)

# tehse are the ASVs within our graph
V(select_igraph)

#these are their node proprieties
selected_nodes<-Generate_node_metrics2(igraph_obj = select_igraph, physeq_obj= network_input_ps)

# define keystones and hubs
keystone_taxa<-KeystoneDetector3(selected_nodes) # 1 kesytone taxa
zi_pi_output<-Zi_Pi_list(select_igraph) # 9 module connectors

# add Zi/Pi to node table
selected_nodes<-left_join(selected_nodes,zi_pi_output[[1]],  by =c("OTU" ="names" ))

#module connectors and hubs
selected_nodes[selected_nodes$Pi>=0.62,]$OTU
selected_nodes[selected_nodes$Zi>=2.5,]






# plot Zi and Pi metrics of module connectivity
Zi_Pi_plot<-ggplot(selected_nodes, aes(x = Pi, y = Zi, color = Class))+
 geom_point(size = selected_nodes$Degree)+
 theme_bw()+
 geom_vline( xintercept = 0.62, linetype = "dashed", colour = "red")+
 geom_hline( yintercept = 2.5, linetype = "dashed", colour = "red")+
 geom_text_repel(data = subset(selected_nodes, Pi >=0.62 | Zi>=2.5),
 aes(x = Pi, y = Zi, label = OTU),
 color = "black")


```





# export network for visualziation in cytoscape

```{r}

#remove NAs as they may crash cytoscape
selected_nodes[is.na(selected_nodes)]<-""

# export nodes
write.csv(x = selected_nodes, 
          file = "./Results/node_prop_l.csv", 
          quote = FALSE,
          row.names = FALSE)

#create edge lsit df
edge_list<-as.data.frame(as_edgelist(igraph_raref_obj_l$bd_cutoff_md2sd))
edge_list$weight<-E(igraph_raref_obj_l$bd_cutoff_md2sd)$weight

#export edge lsit df
write.csv(x = edge_list, 
          file = "./Results/edge_prop_l.csv", 
          quote = FALSE,
          row.names = FALSE)


```


# define correlations of modules with number of flowers

```{r}


library("WGCNA")
eing<-adjustInput_run_eigen_correlation_adjustOutput(
  igraph_obj = igraph_raref_obj_l$bd_cutoff_md2sd,
  phyloseq_obj = prune_taxa(taxa = V(igraph_raref_obj_l$bd_cutoff_md2sd)$name,
                            x = physeq_goodSamples_rarefied)
  )



#plot correlations of modules with emtadata
eigen_corr<-ggplot(data=filter(eing, nodes_in_module>2),aes(x=as.factor(module_number), y=category,fill=ifelse(pval<=0.1,value,NA)))+ # filts the list of objects, drawing one at a time
  geom_tile(colour="#B8B8B8")+
  scale_fill_gradient2("Correlation to \n metadata",guide = "colourbar",high = "#7DEB5F",mid="#F0EE54",low="#F3633F",na.value="white",limits=c(-1,1))+   ylab("")+
  xlab("Cluster/Module")+
  labs(fill="Cluster to Deconstruction")
#  scale_y_discrete(labels=c("leaf dry weight (g)", "MYC2 (fold DD)", "LOX2 (fold DD)"))


```

# keystone and conenctor plots

ASV_13 and ASV_521 (both class Bacilli) are the msot relevant in the network. let's see their abundances

from module 6  is possitively correlated to number of flowers and to plant aboveground biomass."ASV_3", "ASV_30" and "ASV_26" are also predictors of number of flowers

```{r}
  #get boruta stats of ASVs confirmed to be important
key_OTU_ps<-prune_taxa(taxa = c("ASV_13", "ASV_521", "ASV_3", "ASV_30", "ASV_26"), physeq_goodSamples_rarefied)
#turn ps object into a dataframe
melted_ps<-psmelt(key_OTU_ps)


# quick plot bar with relative abudances, for a quick overview of  data
plot_keys_asvs<-
  ggplot(data=melted_ps, aes(x=OTU, y =log10(Abundance), fill=Treatment ))+
  geom_boxplot(aes())+
  theme_bw()+
  theme(axis.title=element_text(size=13, face = "bold"))+
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
  panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))+
  theme(axis.text=element_text(size=10, face = "bold", colour="black"))+
  theme(axis.title.x = element_blank())




```


# explot plot that supports network analysis


```{r}

# make a single pannel for all netowrk support figures
network_support<-ggarrange(keystone_taxa[[1]],
                           Zi_Pi_plot,
                           eigen_corr,
                           plot_keys_asvs,
                           labels = "AUTO")




#export plot
ggsave(network_support,filename = "./Results/network_support.pdf",
      height = 30,
      width = 30,
      units = "cm")

```
